requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
numpy==1.26.4
pydantic==2.7.1
openai==1.40.0
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.6
